# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface Subject {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  category: string;
  /** @nullable */
  icon?: string | null;
  isCustom?: boolean;
  createdAt: string;
}

export interface SubjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  category?: string;
}

export interface Document {
  id: number;
  name: string;
  objectPath: string;
  /** @nullable */
  contentType?: string | null;
  /** @nullable */
  size?: number | null;
  /** @nullable */
  subjectId?: number | null;
  status: string;
  createdAt: string;
}

export interface DocumentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  objectPath: string;
  contentType?: string;
  size?: number;
  subjectId?: number;
}

export interface Question {
  id: number;
  prompt: string;
  options: string[];
  correctIndex: number;
  /** @nullable */
  explanation?: string | null;
  order: number;
}

export type QuizMode = typeof QuizMode[keyof typeof QuizMode];


export const QuizMode = {
  placement: 'placement',
  practice: 'practice',
  exam: 'exam',
} as const;

export interface Quiz {
  id: number;
  title: string;
  mode: QuizMode;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  /** @nullable */
  documentId?: number | null;
  /** @nullable */
  topic?: string | null;
  difficulty: string;
  questionCount: number;
  createdAt: string;
  questions: Question[];
}

export type QuizSummaryMode = typeof QuizSummaryMode[keyof typeof QuizSummaryMode];


export const QuizSummaryMode = {
  placement: 'placement',
  practice: 'practice',
  exam: 'exam',
} as const;

export interface QuizSummary {
  id: number;
  title: string;
  mode: QuizSummaryMode;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  /** @nullable */
  topic?: string | null;
  difficulty: string;
  questionCount: number;
  createdAt: string;
}

export type QuizGenerateInputMode = typeof QuizGenerateInputMode[keyof typeof QuizGenerateInputMode];


export const QuizGenerateInputMode = {
  placement: 'placement',
  practice: 'practice',
  exam: 'exam',
} as const;

export type QuizGenerateInputDifficulty = typeof QuizGenerateInputDifficulty[keyof typeof QuizGenerateInputDifficulty];


export const QuizGenerateInputDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
  mixed: 'mixed',
} as const;

export interface QuizGenerateInput {
  mode: QuizGenerateInputMode;
  subjectId?: number;
  documentId?: number;
  topic?: string;
  title?: string;
  difficulty?: QuizGenerateInputDifficulty;
  /**
     * @minimum 3
     * @maximum 25
     */
  questionCount?: number;
}

export interface AttemptInput {
  answers: number[];
}

export interface QuestionResult {
  questionId: number;
  prompt: string;
  options: string[];
  selectedIndex: number;
  correctIndex: number;
  correct: boolean;
  /** @nullable */
  explanation?: string | null;
}

export interface Attempt {
  id: number;
  quizId: number;
  /** @nullable */
  quizTitle?: string | null;
  /** @nullable */
  mode?: string | null;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  score: number;
  correctCount: number;
  totalQuestions: number;
  level: string;
  /** @nullable */
  feedback?: string | null;
  completedAt: string;
  results: QuestionResult[];
}

export interface AttemptSummary {
  id: number;
  quizId: number;
  /** @nullable */
  quizTitle?: string | null;
  /** @nullable */
  mode?: string | null;
  /** @nullable */
  subjectName?: string | null;
  score: number;
  correctCount: number;
  totalQuestions: number;
  level: string;
  completedAt: string;
}

export interface LearnSection {
  heading: string;
  content: string;
}

export interface LearnSession {
  id: number;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  topic: string;
  title: string;
  summary: string;
  sections: LearnSection[];
  keyPoints: string[];
  nextSteps?: string[];
  createdAt: string;
}

export interface LearnSessionSummary {
  id: number;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  topic: string;
  title: string;
  /** @nullable */
  summary?: string | null;
  createdAt: string;
}

export interface LearnInput {
  /** @minLength 1 */
  topic: string;
  subjectId?: number;
  focus?: string;
}

export interface DashboardSummary {
  totalQuizzes: number;
  totalAttempts: number;
  averageScore: number;
  documentsUploaded: number;
  subjectsStudied: number;
  studyGuides?: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  attempt: 'attempt',
  quiz: 'quiz',
  document: 'document',
  learn: 'learn',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  title: string;
  /** @nullable */
  detail?: string | null;
  timestamp: string;
}

export interface SubjectProgress {
  subjectId: number;
  subjectName: string;
  /** @nullable */
  category?: string | null;
  attempts: number;
  averageScore: number;
  level: string;
}

export type ListQuizzesParams = {
mode?: ListQuizzesMode;
subjectId?: number;
};

export type ListQuizzesMode = typeof ListQuizzesMode[keyof typeof ListQuizzesMode];


export const ListQuizzesMode = {
  placement: 'placement',
  practice: 'practice',
  exam: 'exam',
} as const;