# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List uploaded documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Register an uploaded document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /quizzes/{id}/attempts:
    post:
      operationId: submitAttempt
      tags: [attempts]
      summary: Submit answers for a quiz and get a scored result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttemptInput"
      responses:
        "201":
          description: Scored attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attempt"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Quiz not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /attempts/{id}:
    get:
      operationId: getAttempt
      tags: [attempts]
      summary: Get a full attempt with per-question results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attempt"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /learn/research:
    post:
      operationId: researchTopic
      tags: [learn]
      summary: Generate an AI study guide for a topic
      description: |
        Produces a structured study guide (summary, sections, key points and
        suggested next steps) for a chosen subject or free-form topic.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LearnInput"
      responses:
        "201":
          description: Generated study guide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearnSession"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /learn/sessions/{id}:
    get:
      operationId: getLearnSession
      tags: [learn]
      summary: Get a saved study guide
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Study guide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearnSession"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Document:
      type: object
      required: [id, name, objectPath, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        objectPath:
          type: string
        contentType:
          type: ["string", "null"]
        size:
          type: ["integer", "null"]
        subjectId:
          type: ["integer", "null"]
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    DocumentInput:
      type: object
      required: [name, objectPath]
      properties:
        name:
          type: string
          minLength: 1
        objectPath:
          type: string
          minLength: 1
        contentType:
          type: string
        size:
          type: integer
        subjectId:
          type: integer
...
    AttemptInput:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: integer
    QuestionResult:
      type: object
      required: [questionId, prompt, options, selectedIndex, correctIndex, correct]
      properties:
        questionId:
          type: integer
        prompt:
          type: string
        options:
          type: array
          items:
            type: string
        selectedIndex:
          type: integer
        correctIndex:
          type: integer
        correct:
          type: boolean
        explanation:
          type: ["string", "null"]
    Attempt:
      type: object
      required: [id, quizId, score, correctCount, totalQuestions, level, completedAt, results]
      properties:
        id:
          type: integer
        quizId:
          type: integer
        quizTitle:
          type: ["string", "null"]
        mode:
          type: ["string", "null"]
        subjectId:
          type: ["integer", "null"]
        subjectName:
          type: ["string", "null"]
        score:
          type: number
        correctCount:
          type: integer
        totalQuestions:
          type: integer
        level:
          type: string
        feedback:
          type: ["string", "null"]
        completedAt:
          type: string
          format: date-time
        results:
          type: array
          items:
            $ref: "#/components/schemas/QuestionResult"
...
    LearnSection:
      type: object
      required: [heading, content]
      properties:
        heading:
          type: string
        content:
          type: string
    LearnSession:
      type: object
      required: [id, topic, title, summary, sections, keyPoints, createdAt]
      properties:
        id:
          type: integer
        subjectId:
          type: ["integer", "null"]
        subjectName:
          type: ["string", "null"]
        topic:
          type: string
        title:
          type: string
        summary:
          type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/LearnSection"
        keyPoints:
          type: array
          items:
            type: string
        nextSteps:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
...
    LearnInput:
      type: object
      required: [topic]
      properties:
        topic:
          type: string
          minLength: 1
        subjectId:
          type: integer
        focus:
          type: string
...