# lib/api-zod/src/generated/types/learnSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LearnSection } from './learnSection';

export interface LearnSession {
  id: number;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  topic: string;
  title: string;
  summary: string;
  sections: LearnSection[];
  keyPoints: string[];
  nextSteps?: string[];
  createdAt: Date;
}